\name{mydata}
\alias{mydata}

\docType{data}

\title{mydata} 

\description{
A 6-dimensional design of experiments composed of 50 points uniformly drawn at random over the cube.
}

\usage{data(mydata)}

\format{
  The format is:
List of 6
 $ : num [1:50] 0.12 0.857 0.37 0.68 0.79 ... 
 $ : num [1:50] 0.02 0.945 0.832 0.252 0.593 ...
 $ : num [1:50] 0.2082 0.6732 0.0992 0.9963 0.8108 ...
 $ : num [1:50] 0.488 0.0432 0.3663 0.9466 0.6666 ... 
 $ : num [1:50] 0.5973 0.7092 0.0168 0.4915 0.6477 ... 
 $ : num [1:50] 0.0163 0.2297 0.597 0.0874 0.0682 ... 
}


\examples{
data(mydata)
}

\keyword{datasets}
\keyword{internal}
