% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_feas_vec.R
\name{test_feas_vec}
\alias{test_feas_vec}
\title{Test constraints violation (vectorized)}
\usage{
test_feas_vec(cst, equality = FALSE, tolConstraints = NULL)
}
\arguments{
\item{cst}{matrix of constraints (one column for each constraint function)}

\item{equality}{either FALSE or a Boolean vector defining which constraints are treated as equalities}

\item{tolConstraints}{tolerance (vector) for all constraints. If not provided, set to zero for inequalities and 0.05 for equalities}
}
\value{
A Boolean vector, TRUE if the point if feasible, FALSE if at least one constraint is violated
}
\description{
Test whether a set of constraints are violated or not, depending on their nature (equality or inequality) and tolerance parameters
}
\author{
Mickael Binois 

Victor Picheny
}
